backend/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cron-descriptor==1.2.35
Django==4.2
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-cors-headers==3.14.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
kombu==5.2.4
prompt-toolkit==3.0.38
PyJWT==1.7.1
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.4
six==1.16.0
sqlparse==0.4.4
sudo==1.0.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
